Addressed comments and added support of the external model using Infer

+++ QEfficient/base/modeling_qeff.py
@@ -52,7 +52,7 @@ class QEFFBaseModel(ABC):
         self.onnx_path: Optional[str] = None
         self.qpc_path: Optional[str] = None
         self.qpc_session: Optional[QAICInferenceSession] = None
-
+        model = model.to(torch.float32)
         # Apply the transformations
         any_transformed = False
         for transform in self._pytorch_transforms:
